ugly numbers (previous Google competition)
My rendition...bit ulgy itself in places.  So, as the problem states,
you're having to generate 3^n combinations.  Not possible with straight
nested for-loops.  You either have to repeatedly traverse a while loop
inner or outer...or use some form of recursion...or do what I did where
you put an each loop on a growing set inside of a while loop.